consolidate stm32 mpu util
